Reorganize layout
