+++ src/charge/mod.ts
@@ -1,11 +1,8 @@
-export * from './ch-uri-array-builder.js';
-export * from './ch-uri-array-consumer.js';
-export * from './ch-uri-object-builder.js';
-export * from './ch-uri-object-consumer.js';
 export * from './ch-uri-value-builder.js';
 export * from './ch-uri-value-consumer.js';
 export * from './ch-uri-value.js';
 export * from './formats/mod.js';
+export * from './impl/uri-charge-decoder.js';
 export * from './parse-uri-charge.js';
 export * from './uri-charge-format.js';
 export * from './uri-charge-parser.js';
